app/rag_api/requirements_production.txt
fastapi
uvicorn[standard]
openai
llama-index
llama-index-vector-stores-chroma
llama-index-readers-web
llama-index-readers-file
chromadb>=0.5
python-dotenv
llama-index-embeddings-fastembed
fastembed
psutil
aiohttp
requests
